project/requirements.txt
matplotlib==3.8.0
opacus==1.4.1
pytorch==2.2.2
torchaudion==2.2.2
torchvision==0.17.2
tqdm==4.65.0
